+++ kedro_azureml/cli.py
@@ -2,13 +2,14 @@ import json
 import logging
 import os
 from pathlib import Path
-from typing import List, Optional
+from typing import Optional, Tuple
 
 import click
 from kedro.framework.startup import ProjectMetadata
 
 from kedro_azureml.cli_functions import (
     get_context_and_pipeline,
+    parse_extra_env_params,
     parse_extra_params,
     verify_configuration_directory_for_azure,
     warn_about_ignore_files,
@@ -46,11 +47,11 @@ def commands():
 @click.pass_obj
 @click.pass_context
 def azureml_group(ctx, metadata: ProjectMetadata, env):
-    click.echo(metadata)
     ctx.obj = CliContext(env, metadata)
 
 
 @azureml_group.command()
+@click.argument("subscription_id")
 @click.argument("resource_group")
 @click.argument("workspace_name")
 @click.argument("experiment_name")
@@ -61,6 +62,7 @@ def azureml_group(ctx, metadata: ProjectMetadata, env):
 @click.pass_obj
 def init(
     ctx: CliContext,
+    subscription_id,
     resource_group,
     workspace_name,
     experiment_name,
@@ -75,13 +77,14 @@ def init(
     target_path = Path.cwd().joinpath("conf/base/azureml.yml")
     cfg = CONFIG_TEMPLATE_YAML.format(
         **{
+            "subscription_id": subscription_id,
             "resource_group": resource_group,
             "workspace_name": workspace_name,
             "experiment_name": experiment_name,
             "cluster_name": cluster_name,
-            "environment_name": environment_name,
-            "storage_container": storage_container,
             "storage_account_name": storage_account_name,
+            "storage_container": storage_container,
+            "environment_name": environment_name,
         }
     )
     target_path.write_text(cfg)
@@ -147,6 +150,12 @@ def init(
     help="Parameters override in form of JSON string",
 )
 @click.option("--wait-for-completion", type=bool, is_flag=True, default=False)
+@click.option(
+    "--env-var",
+    type=str,
+    multiple=True,
+    help="Environment variables to be injected in the steps, format: KEY=VALUE",
+)
 @click.pass_obj
 @click.pass_context
 def run(
@@ -158,14 +167,15 @@ def run(
     pipeline: str,
     params: str,
     wait_for_completion: bool,
+    env_var: Tuple[str],
 ):
     """Runs the specified pipeline in Azure ML Pipelines; Additional parameters can be passed from command line.
     Can be used with --wait-for-completion param to block the caller until the pipeline finishes in Azure ML.
     """
     params = json.dumps(p) if (p := parse_extra_params(params)) else ""
-    assert (
-        subscription_id
-    ), f"Please provide Azure Subscription ID or set `{AZURE_SUBSCRIPTION_ID}` env"
+
+    if subscription_id:
+        click.echo(f"Overriding Azure Subscription ID for run to: {subscription_id}")
 
     if aml_env:
         click.echo(f"Overriding Azure ML Environment for run by: {aml_env}")
@@ -175,7 +185,8 @@ def run(
     verify_configuration_directory_for_azure(click_context, ctx)
 
     mgr: KedroContextManager
-    with get_context_and_pipeline(ctx, image, pipeline, params, aml_env) as (
+    extra_env = parse_extra_env_params(env_var)
+    with get_context_and_pipeline(ctx, image, pipeline, params, aml_env, extra_env) as (
         mgr,
         az_pipeline,
     ):
@@ -287,7 +298,7 @@ def compile(
 )
 @click.pass_obj
 def execute(
-    ctx: CliContext, pipeline: str, node: str, params: str, azure_outputs: List[str]
+    ctx: CliContext, pipeline: str, node: str, params: str, azure_outputs: Tuple[str]
 ):
     # 1. Run kedro
     parameters = parse_extra_params(params)