+++ tests/test_generator.py
@@ -120,3 +120,21 @@ def test_get_target_resource_from_node_tags_raises_exception(
         )
         with pytest.raises(ConfigException):
             generator.get_target_resource_from_node_tags(node)
+
+
+def test_azure_pipeline_with_custom_env_vars(dummy_plugin_config, dummy_pipeline):
+    pipeline_name = "unit_test_pipeline"
+    node = MagicMock()
+    node.tags = ["compute-2", "compute-3"]
+    for t in node.tags:
+        dummy_plugin_config.azure.compute[t] = ComputeConfig(**{"cluster_name": t})
+    with patch.dict(
+        "kedro.framework.project.pipelines", {pipeline_name: dummy_pipeline}
+    ):
+        generator = AzureMLPipelineGenerator(
+            pipeline_name, "local", dummy_plugin_config, {}, extra_env={"ABC": "def"}
+        )
+
+        for node in generator.generate().jobs.values():
+            assert "ABC" in node.environment_variables
+            assert node.environment_variables["ABC"] == "def"